Remove render_and_update_links as default.
This doesn't actaully affirm that we re-render the page
even if nothing has changed, this code path can only be
followed if it's already going to make a new revision for
sure.

I don't think there are any paths which have
render_and_update_links not called, but that's because
other field changes require it, not because it's unconditional.

We'll figure out how to do "rerender if edit is empty" later
when this stuff is a bit more settled.